Added user authorization

+++ src/tasks/tasks.service.ts
@@ -6,16 +6,18 @@ import { GetTasksFilterDto } from './dto/get-tasks-filter.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
 import { Repository } from 'typeorm';
+import { User } from '../auth/user.entity';
 
 @Injectable()
 export class TasksService {
   constructor(@InjectRepository(Task) private tasksRepository: Repository<Task>) { }
 
 
-  async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
+  async getTasks(filterDto: GetTasksFilterDto, user: User): Promise<Task[]> {
     const { search, status } = filterDto;
 
     const query = this.tasksRepository.createQueryBuilder('task');
+    query.where('task.userId = :userId', { userId: user.id })
 
     if (status) {
       query.andWhere('task.status= :status', { status })
@@ -32,8 +34,8 @@ export class TasksService {
     return tasks;
   }
 
-  async getTaskById(id: string): Promise<Task> {
-    const foundTask = await this.tasksRepository.findOne({ where: { id } })
+  async getTaskById(id: string, user: User): Promise<Task> {
+    const foundTask = await this.tasksRepository.findOne({ where: { id, userId: user.id } })
     if (!foundTask) {
       throw new NotFoundException(`Task with ID: '${id}' was not found`)
     }
@@ -42,12 +44,13 @@ export class TasksService {
   }
 
 
-  async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
+  async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
     const { title, description } = createTaskDto;
     const task = this.tasksRepository.create({
       title,
       description,
-      status: TaskStatus.DONE
+      status: TaskStatus.DONE,
+      user
     })
 
     await this.tasksRepository.save(task)
@@ -55,15 +58,15 @@ export class TasksService {
     return task;
   }
 
-  async updateTaskStatus(updateTaskDto: UpdateTaskDto): Promise<Task> {
+  async updateTaskStatus(updateTaskDto: UpdateTaskDto, user: User): Promise<Task> {
     const { id, status } = updateTaskDto;
-    let task = await this.getTaskById(id);
+    let task = await this.getTaskById(id, user);
     task.status = status['status'];
     return task;
   }
 
-  async deleteTask(id: string): Promise<void> {
-    const result = await this.tasksRepository.delete(id)
+  async deleteTask(id: string, user: User): Promise<void> {
+    const result = await this.tasksRepository.delete({ id, userId: user.id })
     if (result.affected === 0) {
       throw new NotFoundException(`Task with ID: '${id}' was not found`)
     }
